Corrected tests 22 and 23.
test22.s would previously read the UART status flag once,
then entered a loop where the flag value is checked to see
if the UART has some input ready for the program. However, this loop
did not refresh the flag value, which meant if the first flag read
didn't signal that some input was ready, the loop would be inifinate.
This change correctly refreshes the flag value in every loop.

test23.s had a loop that has the same fundamental problem as with test22.
However, pasim's UART implementation always accepts output to the UART
from the program, meaning the UART status flag is always the same.
This change updates the test to just check the flag once, and then doesn't
output '!' if the flag is not signalling that the UART is ready for output.